+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 0.1.3
+- fixed issue when exporting properties with false (checkbox) value
+
 # 0.1.2
 - added product custom fields to product export
 - fixed coupon error handling for ineligible rules